complete 128 bit hash for string to number for account mapping and fees

+++ fio.contracts/contracts/fio.fee/fio.fee.hpp
@@ -29,14 +29,14 @@ namespace fioio {
     struct [[eosio::action]] fiofee {        // FIO fee
         uint64_t fee_id;     // one up index starting at 0
         string end_point;  // the blockchain endpoint to which the fee relates.
-        uint64_t end_point_hash;    // this is a hash of the end point, lending search-ability
+        uint128_t end_point_hash;    // this is a hash of the end point, lending search-ability
         uint64_t type;      // this is the fee type from the feetype enumeration.
         uint64_t suf_amount;      // this is the amount of the fee in small units of FIO.
         // 1 billion per fio as of 04/23/2019
 
         uint64_t primary_key() const { return fee_id; }
 
-        uint64_t by_endpoint() const { return end_point_hash; }
+        uint128_t by_endpoint() const { return end_point_hash; }
 
         uint64_t by_type() const { return type; }
 
@@ -46,7 +46,7 @@ namespace fioio {
 
     // FIO fee table
     typedef multi_index<"fiofees"_n, fiofee,
-            indexed_by<"byendpoint"_n, const_mem_fun < fiofee, uint64_t, &fiofee::by_endpoint>>,
+            indexed_by<"byendpoint"_n, const_mem_fun < fiofee, uint128_t, &fiofee::by_endpoint>>,
     indexed_by<"bytype"_n, const_mem_fun<fiofee, uint64_t, &fiofee::by_type>
     >>
     fiofee_table;
